🌠 Tweak list shrinker so it doesn't fall over on lists of higher lengths
Previously if you tried to find the minimal list of at least length 20, it would never complete. This is because none of the shrinks satisfied the predicate, but the list had sooooo many shrinks. This is because it was edging towards the current list whilst generation combinations. So a list of length 20 would generate lengths of [0, 10, 15, 18, 19] and do n choose k for each length. For reference, 20 choose 10 = 200k combinations.

Now, rather than exhausting combinations, we bisect the list. If the a failing combination is severed by that bisection, then we fallback by dropping a single element. Eventually, either the bisection won't severe the failing combo, and we'll make a great stride towards the minimal shrink, or, worst case scenario, we'll drop enough elements one-by-one so that we brute force our way there (in an achievable way).