🌍 #6 Implement void properties with no configuration
Most of the invocation logic layed out in #8 can probably be implemented as part of the core framework.

That way it will be able to be utilized by future integrations.

For example, the idea of a property test case being able to return void, bool, or another property. Also preconditions.