🌍 #86 Add an optimized Int32 example space
We can take advantage of contextualized shrinkers, so that we can smartly avoid shrinks that we know we have convered.

For example, if we're shrinking from 10 towards 0, and the real smallest value is 7, we used to do:

10 -> [shrink from 10 to 0] -> 0 -> 5 -> 8 -> [shrink from 8 to 0] -> 4 -> 6 -> 7

Now the example space will be traversed as:

10 -> [shrink from 10 to 0] -> 0 -> 5 -> 8 -> [shrink from 8 to 5] -> 7

At the second level of shrinking (when we shrink 8), we know that the value must be greater than 5, as we already traversed over 5 - therefore we can rule out 4. 6 also happens to be dropped in this scenario because the shrinking algorithm edges from 5 towards 8, which happens to skip 6.