🌍 #86 Add MinimumWithMetrics, which reports number of shrinks (#91)
Create `ContextualShrinker` abstraction, and use in tests to control the number of shrinks that an arbitrary generator might take. This is super handy for tests, and could improve a lot of previously written tests. In the past we dodgely did this kind of thing with impure functions. Now we have a nice context that gets passed to stick that sort of information on.