🌍 #126 Add Gen.Choose (#127)
I thought this generator would leverage an unbiased int32 gen, but I was wrong. It's better to start from primitives (because we don't want any of the features of the int32 gen, and we want custom shrinking).

It's important that we generate the index with the weighting, then we shrink the index without the weighting. Otherwise, shrinking the weighted index might end up at the same generator.