🌍 #24 Make `Test` more simple by replacing the Func with a lazy output
This means that consumers of this interface don't need to orchestrate the output by calling the test's function with the test's input - which therefore eliminates the bad state of mixing test functions and test inputs up.

This breaks reporting on nested properties, but simplifies the next tidy up, which should fix that.