🌍 #24 Delete IGenInjectionConfigurationFilter API, and make the GenProviderAttribute the single point of composition.
This enables super-generic customisation via attributes - the attribute merely defines the generator, which gives the flexibility of using any gen API, rather than having to slot in between builder configurations (and losing specific types if you wanted to do NoShrink(), for example).

Also, GenAttribute has much better validation. It ensures that the type of the IGen conforms to the type of the parameter.

This meant that some existing out-of-the-box GenAttributes could not be migrated, as they are not much use alone (WithBias, ShrinkTowards)