🌍 #210 Removes Properties, and previous Xunit Factory configuration. Adds GenFactoryAttribute. Make IGenFactory implementation immutable.
This API change brings much more type-safety to configuring the factory.

It also provides a similar experience to GenProviderAttribute. You can now easily re-use the same factory in the attribute and elsewhere.

IGenFactory immutability brings this API inline with all the other GalaxyCheck APIs in terms of immutability. Hopefully this avoids some expectations being subverted. It also fixes a bug that existed in Gen.Parameters.