🌍 #214 Improve ExampleSpace machinery for integers
Whilst adding tests for Int64, I found that shrinking was broken in some cases. This was because of the encountered value tracking that exists in TraverseUnencountered. Because it's based on 32-bit hashcodes, it was possible for 64-bit integers to clash, meaning false-positives on encounters (and erroneously skipping unencountered values).

It turns out that this code is unnecessary, since we added the ContexualShrinker API. The ContextualShrinkers for Int32 and Int64 skip successfully based on actual value, not based on a hashcode. They also offer a much richer API for skipping unnecessary shrinks.

Therefore, we can rip out the old algorithm for skipping encountered values. There is much code to tidy up after this.