🌍 #361 Generator stability V2
This was initially trialled in #337, but was rolled back as it adversely affected randomness.

Second attempt is much cleaner and tidier, doesn't affect randomness, and is opt-in (a snaphotting impl would choose to turn it on).

You can call `SetRngWaypoint` on your root generator, then any generators that it's comprised of can generate stable seeds for themselves using `ReferenceRngWaypoint`.