🌍 #376 Refactor - combine the sync and async check state machines
This was a four month long battle. I picked it up many times throughout, each time being tempted by `git branch -D terra/async-refactor`.

Previously, one of the core pieces of infrastructure in GC was copy+pasted between Check/CheckAsync, it was a lot of lines of code to keep in sync. Even though CheckAsync is the primary usecase (it's the entrypoint used by GalaxyCheck.Xunit), I was not keen to drop the synchronous overload, as it would be a giant pain to litter the codebase with async/await.

This also means we can mostly forget about testing the async version of the code, as it barely deviates from the sync version at all, only at the entrypoint in `CheckAsync`.