🌍 #371 Fix infinite rendering loop on static properties
It's not meaningful to render static properties on a (non-static) counterexample, so don't.

Also, this was causing an infinite loop when rendering a counterexample containing System.Drawing.Color. I've put in an extra safeguard, so that the object renderer checks that it isn't rendering a type that it hasn't just seen.