🚧 Rename `NebulaCheck` => `GalaxyCheck.Stable`
Due to the .NET framework, we need to synthesize a different assembly so that GalaxyCheck can be used to test itself.

The intention of this rename is to make this clear, and to expand the available samples of GalaxyCheck by making it more obvious that it's being used in its own test suite.